+++ milkman/src/test/java/milkman/ui/commands/EnvironmentTemplaterTest.java
@@ -1,12 +1,14 @@
 package milkman.ui.commands;
 
 
-import milkman.ctrl.EnvironmentTemplater;
+import milkman.templater.EnvironmentTemplater;
 import milkman.domain.Environment;
 import milkman.domain.Environment.EnvironmentEntry;
+import milkman.templater.PrefixedTemplaterResolver;
 import org.junit.jupiter.api.Test;
 
 import java.util.Collections;
+import java.util.List;
 import java.util.Optional;
 import java.util.UUID;
 
@@ -22,8 +24,9 @@ class EnvironmentTemplaterTest {
 		globalEnvironment.getEntries().add(new EnvironmentEntry(UUID.randomUUID().toString(),"test2", "value2", true));
 		globalEnvironment.getEntries().add(new EnvironmentEntry(UUID.randomUUID().toString(),"test4", "value4", false));
 		globalEnvironment.getEntries().add(new EnvironmentEntry(UUID.randomUUID().toString(),"test6", "{{test2}}", true));
-		
-		EnvironmentTemplater sut = new EnvironmentTemplater(Optional.empty(), Collections.singletonList(globalEnvironment));
+
+		PrefixedTemplaterResolver res = new PrefixedTemplaterResolver(List.of());
+		EnvironmentTemplater sut = new EnvironmentTemplater(Optional.empty(), Collections.singletonList(globalEnvironment), res);
 		String output = sut.replaceTags("test1 {{test2}} {{test3}} {{test4}} test5 {{test6}} someTail");
 		assertThat(output).isEqualTo("test1 value2 {{test3}} {{test4}} test5 value2 someTail");